third_party/WebKit/LayoutTests/media/video-playing-and-pause.html

<!DOCTYPE HTML>
<html>
<body>
<video controls></video>
<p>Test that pausing the media element in "playing" event handler pauses the media immediately.</p>
<script src=media-file.js></script>
<script src=video-test.js></script>
<script>
    video.src = findMediaFile("video", "content/test.mp4");
    waitForEvent("timeupdate");
    waitForEvent('play', function () { video.pause(); } );
    waitForEvent('pause', function () { testExpected("video.paused", true); endTest(); } );

    run("video.play()");
    
    consoleWrite("SCRIPT DONE");
</script>
</body>
</html>
